requirements.txt
apache-airflow==1.10.11
boto3==1.14.23
SQLAlchemy==1.3.18
pandas==1.1.1
